requirements.txt
Django
psycopg2
GeoAlchemy2
django-geojson-serializer
djangorestframework
shapely